Refactor: Dialog 컴포넌트 분리
