+++ src/pages/PostsManagerPage.tsx
@@ -3,16 +3,8 @@ import { Edit2, MessageSquare, Plus, Search, ThumbsDown, ThumbsUp, Trash2 } from
 import { useLocation, useNavigate } from "react-router-dom"
 import { Button } from "../shared/ui/button/Button"
 import { Card, CardContent, CardHeader, CardTitle } from "../shared/ui/card/Card"
+import { Dialog, DialogContent, DialogHeader, DialogTitle } from "../shared/ui/dialog/Dialog"
 import {
-  // Button,
-  // Card,
-  // CardContent,
-  // CardHeader,
-  // CardTitle,
-  Dialog,
-  DialogContent,
-  DialogHeader,
-  DialogTitle,
   Input,
   Select,
   SelectContent,
